hw06/requirements.txt
python-telegram-bot
numpy~=1.18.3